requirements.txt
python-dotenv==1.1.1
requests==2.32.5
flask==3.0.0
